move misc to main
