+++ packages/insomnia/src/main/ipc/main.ts
@@ -8,10 +8,10 @@ import { app, BrowserWindow, ipcMain, IpcRendererEvent, shell } from 'electron';
 import fs from 'fs';
 
 import { axiosRequest } from '../../network/axios-request';
-import { authorizeUserInWindow } from '../../network/o-auth-2/misc';
 import { SegmentEvent, trackPageView, trackSegmentEvent } from '../analytics';
 import { exportAllWorkspaces } from '../export';
 import installPlugin from '../install-plugin';
+import { authorizeUserInWindow } from '../misc';
 import { cancelCurlRequest, curlRequest } from '../network/libcurl-promise';
 import { WebSocketBridgeAPI } from '../network/websocket';
 import { gRPCBridgeAPI } from './grpc';
